fix:update Double to Float

+++ spring-ai-alibaba-example-nacos/rag/agentic-rag-system/src/main/java/com/example/rag/service/AzureOpenAIService.java
@@ -54,7 +54,7 @@ public class AzureOpenAIService {
                 .toFuture()
                 .thenApply(embeddings -> {
                     if (embeddings.getData() != null && !embeddings.getData().isEmpty()) {
-                        List<Double> embedding = embeddings.getData().get(0).getEmbedding();
+                        List<Float> embedding = embeddings.getData().get(0).getEmbedding();
                         return embedding.toString();
                     }
                     return "";
